Fix stripping export statements on non-build files
We convert named exports to a default namespace export for better
interop with VSCode intellisense on built files.

Doing this was causing problems when importing from files that are not
build targets, since the expected named export statements were going
missing. This isn't a problem for built files, because we apply the
syntheticnamedexports setting to them.

This issue is resolved by simply skipping the stripping process for
these non-built files, since we never have to care about intellisense
for a non-existent build artefact.